requirements.txt
# ─── Web framework ────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==1.10.21

# ─── Auth ─────────────────────────────────────────────────────────────────────
PyJWT==2.8.0

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy==1.4.49
alembic==1.13.1
psycopg2-binary==2.9.9

# ─── AI / LLM ─────────────────────────────────────────────────────────────────
google-generativeai==0.5.4

# ─── Embeddings & NLP (Python 3.12 + Windows compatible) ──────────────────────
torch==2.2.2
transformers==4.37.2
sentence-transformers==2.3.1

# ─── ML / Math ────────────────────────────────────────────────────────────────
scikit-learn>=1.3.0
numpy==1.26.4

# ─── Document ingestion ───────────────────────────────────────────────────────
pdfplumber==0.9.0

# ─── Streamlit UI ─────────────────────────────────────────────────────────────
streamlit==1.30.0

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
langdetect==1.0.9
openai==1.12.0